# api_specification.json
{"openapi":"3.1.0","info":{"title":"Memora API","description":"AI-powered personal memory assistant API","version":"0.1.0"},"paths":{"/extract_and_save":{"post":{"summary":"Extract And Save","description":"Extract content from a URL and save it in the database.","operationId":"extract_and_save_extract_and_save_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExtractRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MemoraItem"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/search":{"get":{"summary":"Search","description":"Search for saved content using a natural language query.\n\nOptionally filter by content type and platform.","operationId":"search_search_get","parameters":[{"name":"user_id","in":"query","required":true,"schema":{"type":"string","title":"User Id"}},{"name":"query","in":"query","required":true,"schema":{"type":"string","title":"Query"}},{"name":"top_k","in":"query","required":false,"schema":{"anyOf":[{"type":"integer","maximum":20,"minimum":1},{"type":"null"}],"default":5,"title":"Top K"}},{"name":"content_type","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Content Type"}},{"name":"platform","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Platform"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MemoraItem"},"title":"Response Search Search Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"summary":"Health Check","description":"Health check endpoint.","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"ExtractRequest":{"properties":{"user_id":{"type":"string","title":"User Id","description":"User ID"},"url":{"type":"string","maxLength":2083,"minLength":1,"format":"uri","title":"Url","description":"URL to extract content from"}},"type":"object","required":["user_id","url"],"title":"ExtractRequest","description":"Request for extracting content from a URL and saving it."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MemoraItem":{"properties":{"id":{"type":"string","title":"Id"},"user_id":{"type":"string","title":"User Id"},"url":{"type":"string","title":"Url"},"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description"},"tags":{"items":{"type":"string"},"type":"array","title":"Tags"},"timestamp":{"type":"string","format":"date-time","title":"Timestamp"},"content_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Content Type"},"platform":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Platform"},"similarity_score":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Similarity Score"}},"type":"object","required":["id","user_id","url","title","description","tags","timestamp"],"title":"MemoraItem","description":"Item stored in the Memora database."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}